+++ packages/moonstone/Panels/ApplicationCloseButton.js
@@ -24,13 +24,13 @@ const ApplicationCloseButton = kind({
 	propTypes: /** @lends moonstone/Panels.ApplicationCloseButton.prototype */ {
 		/**
 		 * The background-color opacity of this button; valid values are `'opaque'`, `'translucent'`,
-		 * and `'transparent'`.
+		 * `'lightTranslucent'` and `'transparent'`.
 		 *
 		 * @type {String}
 		 * @default 'transparent'
 		 * @public
 		 */
-		backgroundOpacity: PropTypes.oneOf(['opaque', 'translucent', 'transparent']),
+		backgroundOpacity: PropTypes.oneOf(['opaque', 'translucent', 'lightTranslucent', 'transparent']),
 		/**
 		 * A function to run when app close button is clicked
 		 *